+++ packages/text-area/src/index.ts
@@ -1,0 +1,1 @@
+export { State, default } from './TextArea';