+++ packages/text-input/src/TextInput.tsx
@@ -4,16 +4,16 @@ import { css, cx } from '@leafygreen-ui/emotion';
 import CheckmarkIcon from '@leafygreen-ui/icon/dist/Checkmark';
 import CheckmarkWithCircleIcon from '@leafygreen-ui/icon/dist/CheckmarkWithCircle';
 import WarningIcon from '@leafygreen-ui/icon/dist/Warning';
+import InteractionRing from '@leafygreen-ui/interaction-ring';
 import { uiColors } from '@leafygreen-ui/palette';
-import { useUsingKeyboardContext } from '@leafygreen-ui/leafygreen-provider';
 import {
   createDataProp,
   HTMLElementProps,
   Either,
   IdAllocator,
 } from '@leafygreen-ui/lib';
+import { Description, Label } from '@leafygreen-ui/typography';
 
-const inputSelectorProp = createDataProp('input-selector');
 const iconSelectorProp = createDataProp('icon-selector');
 
 export const State = {
@@ -42,7 +42,7 @@ const Mode = {
 
 type Mode = typeof Mode[keyof typeof Mode];
 
-interface BaseTextInputProps extends HTMLElementProps<'input'> {
+interface TextInputProps extends HTMLElementProps<'input', HTMLInputElement> {
   /**
    * id associated with the TextInput component.
    */
@@ -51,7 +51,7 @@ interface BaseTextInputProps extends HTMLElementProps<'input'> {
   /**
    * Text shown in bold above the input element.
    */
-  label?: string;
+  label?: string | null;
 
   /**
    * Text that gives more detail about the requirements for the input.
@@ -106,42 +106,20 @@ interface BaseTextInputProps extends HTMLElementProps<'input'> {
   darkMode?: boolean;
 
   type?: TextInputType;
+
+  ['aria-labelledby']?: string;
 }
 
 type AriaLabels = 'label' | 'aria-labelledby';
-type TextInputProps = Either<BaseTextInputProps, AriaLabels>;
+type AccessibleTextInputProps = Either<TextInputProps, AriaLabels>;
 
-const interactionRing = css`
-  transition: all 150ms ease-in-out;
-  transform: scale(0.9, 0.8);
-  border-radius: 7px;
-  position: absolute;
-  top: -3px;
-  bottom: -3px;
-  left: -3px;
-  right: -3px;
-  pointer-events: none;
+const interactionRingStyle = css`
+  width: 100%;
 `;
 
 const textInputStyle = css`
   display: flex;
   flex-direction: column;
-`;
-
-const labelStyle = css`
-  font-size: 14px;
-  font-weight: bold;
-  line-height: 16px;
-  padding-bottom: 4px;
-`;
-
-const descriptionStyle = css`
-  font-size: 14px;
-  line-height: 16px;
-  font-weight: normal;
-  padding-bottom: 4px;
-  margin-top: 0px;
-  margin-bottom: 0px;
 `;
 
 const inputContainerStyle = css`
@@ -161,15 +139,15 @@ const inputStyle = css`
   font-family: Akzidenz, ‘Helvetica Neue’, Helvetica, Arial, sans-serif;
   border: 1px solid;
   z-index: 1;
+  outline: none;
 
   &::placeholder {
     color: ${uiColors.gray.base};
   }
 
   &:focus {
-    outline: none;
     z-index: 2;
-    border-color: #9dd0e7;
+    border-color: ${uiColors.blue.light1};
     transition: border-color 150ms ease-in-out;
 
     & ~ ${iconSelectorProp.selector} {
@@ -179,19 +157,6 @@ const inputStyle = css`
 
   &:disabled {
     cursor: not-allowed;
-  }
-`;
-
-const interactionRingFocusStyle = css`
-  ${inputSelectorProp.selector}:focus ~ & {
-    transform: scale(1);
-    z-index: 1;
-  }
-`;
-
-const interactionRingHoverStyle = css`
-  ${inputSelectorProp.selector}:hover ~ & {
-    transform: scale(1);
   }
 `;
 
@@ -222,11 +187,6 @@ const errorMessageStyle = css`
 `;
 
 interface ColorSets {
-  interactionRing: string;
-  interactionRingFocus: string;
-  labelColor: string;
-  disabledLabelColor: string;
-  descriptionColor: string;
   inputColor: string;
   inputBackgroundColor: string;
   disabledColor: string;
@@ -241,11 +201,6 @@ interface ColorSets {
 
 const colorSets: Record<Mode, ColorSets> = {
   [Mode.Light]: {
-    interactionRing: uiColors.gray.light2,
-    interactionRingFocus: '#9dd0e7',
-    labelColor: uiColors.gray.dark2,
-    disabledLabelColor: uiColors.gray.dark1,
-    descriptionColor: uiColors.gray.dark1,
     inputColor: uiColors.gray.dark3,
     inputBackgroundColor: uiColors.white,
     disabledColor: uiColors.gray.base,
@@ -258,11 +213,6 @@ const colorSets: Record<Mode, ColorSets> = {
     validBorder: uiColors.green.base,
   },
   [Mode.Dark]: {
-    interactionRing: uiColors.gray.dark1,
-    interactionRingFocus: uiColors.blue.base,
-    labelColor: uiColors.white,
-    disabledLabelColor: uiColors.gray.light1,
-    descriptionColor: uiColors.gray.light1,
     inputColor: uiColors.white,
     inputBackgroundColor: '#394F5A',
     disabledColor: uiColors.gray.dark1,
@@ -331,7 +281,9 @@ const idAllocator = IdAllocator.create('text-input');
  * @param props.className className supplied to the TextInput container.
  * @param props.darkMode determines whether or not the component appears in dark mode.
  */
-const TextInput = React.forwardRef(
+const TextInput: React.ComponentType<React.PropsWithRef<
+  AccessibleTextInputProps
+>> = React.forwardRef(
   (
     {
       label,
@@ -349,14 +301,13 @@ const TextInput = React.forwardRef(
       className,
       darkMode = false,
       ...rest
-    }: TextInputProps,
+    }: AccessibleTextInputProps,
     forwardRef: React.Ref<HTMLInputElement>,
   ) => {
     const mode = darkMode ? Mode.Dark : Mode.Light;
     const isControlled = typeof controlledValue === 'string';
     const [uncontrolledValue, setValue] = useState('');
     const value = isControlled ? controlledValue : uncontrolledValue;
-    const { usingKeyboard: showFocus } = useUsingKeyboardContext();
     const id = useMemo(() => propsId ?? idAllocator.generate(), [propsId]);
 
     function onValueChange(e: React.ChangeEvent<HTMLInputElement>) {
@@ -382,63 +333,50 @@ const TextInput = React.forwardRef(
     return (
       <div className={cx(textInputStyle, className)}>
         {label && (
-          <label
-            htmlFor={id}
-            className={cx(
-              labelStyle,
-              css`
-                color: ${disabled
-                  ? colorSets[mode].disabledLabelColor
-                  : colorSets[mode].labelColor};
-              `,
-            )}
-          >
+          <Label darkMode={darkMode} htmlFor={id} disabled={disabled}>
             {label}
-          </label>
+          </Label>
         )}
         {description && (
-          <p
-            className={cx(
-              descriptionStyle,
-              css`
-                color: ${colorSets[mode].descriptionColor};
-              `,
-            )}
-          >
-            {description}
-          </p>
+          <Description darkMode={darkMode}>{description}</Description>
         )}
         <div className={inputContainerStyle}>
-          <input
-            {...inputSelectorProp.prop}
-            {...rest}
-            aria-labelledby={ariaLabelledBy}
-            type={type}
-            className={cx(
-              inputStyle,
-              css`
-                color: ${colorSets[mode].inputColor};
-                background-color: ${colorSets[mode].inputBackgroundColor};
-
-                &:focus {
-                  border: 1px solid ${colorSets[mode].inputBackgroundColor};
-                }
-
-                &:disabled {
-                  color: ${colorSets[mode].disabledColor};
-                  background-color: ${colorSets[mode].disabledBackgroundColor};
-                }
-              `,
-              { [getStatefulInputStyles(state, optional, mode)]: !disabled },
-            )}
-            value={value}
-            required={!optional}
+          <InteractionRing
+            className={interactionRingStyle}
+            darkMode={darkMode}
             disabled={disabled}
-            placeholder={placeholder}
-            onChange={onValueChange}
-            ref={forwardRef}
-            id={id}
-          />
+          >
+            <input
+              {...rest}
+              aria-labelledby={ariaLabelledBy}
+              type={type}
+              className={cx(
+                inputStyle,
+                css`
+                  color: ${colorSets[mode].inputColor};
+                  background-color: ${colorSets[mode].inputBackgroundColor};
+
+                  &:focus {
+                    border: 1px solid ${colorSets[mode].inputBackgroundColor};
+                  }
+
+                  &:disabled {
+                    color: ${colorSets[mode].disabledColor};
+                    background-color: ${colorSets[mode]
+                      .disabledBackgroundColor};
+                  }
+                `,
+                { [getStatefulInputStyles(state, optional, mode)]: !disabled },
+              )}
+              value={value}
+              required={!optional}
+              disabled={disabled}
+              placeholder={placeholder}
+              onChange={onValueChange}
+              ref={forwardRef}
+              id={id}
+            />
+          </InteractionRing>
           {!disabled && (
             <div {...iconSelectorProp.prop} className={inputIconStyle}>
               {state === State.Valid && (
@@ -466,25 +404,6 @@ const TextInput = React.forwardRef(
                 </div>
               )}
             </div>
-          )}
-          {!disabled && (
-            <div
-              className={cx(
-                interactionRing,
-                css`
-                  background-color: ${colorSets[mode].interactionRing};
-                `,
-                interactionRingHoverStyle,
-                {
-                  [interactionRingFocusStyle]: showFocus,
-                  [css`
-                    ${inputSelectorProp.selector}:focus ~ & {
-                      background-color: ${colorSets[mode].interactionRingFocus};
-                    }
-                  `]: showFocus,
-                },
-              )}
-            />
           )}
         </div>
         {!disabled && state === State.Error && errorMessage && (