+++ website/components/CodeDocs.tsx
@@ -6,6 +6,7 @@ import ActivityFeedIcon from '@leafygreen-ui/icon/dist/ActivityFeed';
 import Button from '@leafygreen-ui/button';
 import Card from '@leafygreen-ui/card';
 import Code from '@leafygreen-ui/code';
+import Copyable from '@leafygreen-ui/copyable';
 import Modal from '@leafygreen-ui/modal';
 import { Tabs, Tab } from '@leafygreen-ui/tabs';
 import { Subtitle, Body } from '@leafygreen-ui/typography';
@@ -15,6 +16,7 @@ import { useViewportSize } from '@leafygreen-ui/hooks';
 import { BaseLayoutProps } from 'utils/types';
 import { GridContainer, GridItem } from 'components/Grid';
 import PropTable, { ReadmeMarkdown } from 'components/PropTable';
+import TypeDefinition from 'components/TypeDefinition';
 
 const topAlignment = css`
   margin-top: ${spacing[4]}px;
@@ -46,15 +48,30 @@ const mobileInstallMargin = css`
   margin-top: 50px;
 `;
 
+const changelogStyles = css`
+  color: ${uiColors.gray.dark3};
+  pointer-events: none;
+
+  & > h2 {
+    padding-top: ${spacing[3]}px;
+    border-top: 1px solid ${uiColors.gray.light2};
+  }
+
+  a {
+    color: ${uiColors.gray.dark3};
+    text-decoration: none;
+  }
+`;
+
 interface VersionCardProps {
-  version: string;
+  version?: string;
   changelog: string;
   isMobile?: boolean;
 }
 
 interface InstallProps {
   component: string;
-  version: string;
+  version?: string;
   changelog: string;
 }
 
@@ -67,6 +84,7 @@ function VersionCard({
 
   return (
     <Card className={cx(topAlignment, versionCard)}>
+      {/* TODO: Provide fallback if no version */}
       <Subtitle className={subtitlePadding}>Version {version}</Subtitle>
       <Button
         size={isMobile ? 'large' : 'normal'}
@@ -82,10 +100,7 @@ function VersionCard({
       </Button>
       <Modal open={openModal} setOpen={setOpenModal}>
         <div
-          className={css`
-            text-decoration: none;
-            color: ${uiColors.gray.dark3};
-          `}
+          className={changelogStyles}
           dangerouslySetInnerHTML={{ __html: changelog }}
         ></div>
       </Modal>
@@ -102,11 +117,11 @@ function MobileInstall({ component, version, changelog }: InstallProps) {
           <Body weight="medium" className={mt3}>
             Yarn
           </Body>
-          <Code language="js">{`yarn add @leafygreen-ui/${component}`}</Code>
+          <Copyable>{`yarn add @leafygreen-ui/${component}`}</Copyable>
           <Body weight="medium" className={mt3}>
             NPM
           </Body>
-          <Code language="js">{`npm install @leafygreen-ui/${component}`}</Code>
+          <Copyable>{`npm install @leafygreen-ui/${component}`}</Copyable>
         </div>
       </GridItem>
       <GridItem sm={12}>
@@ -134,7 +149,7 @@ function DesktopInstall({ component, changelog, version }: InstallProps) {
             <Body weight="medium" className={mb1}>
               Yarn
             </Body>
-            <Code language="js">{`yarn add @leafygreen-ui/${component}`}</Code>
+            <Copyable>{`yarn add @leafygreen-ui/${component}`}</Copyable>
           </div>
         </GridItem>
         <GridItem md={5} lg={5}>
@@ -146,7 +161,7 @@ function DesktopInstall({ component, changelog, version }: InstallProps) {
           <Body weight="medium" className={mb1}>
             NPM
           </Body>
-          <Code language="js">{`npm install @leafygreen-ui/${component}`}</Code>
+          <Copyable>{`npm install @leafygreen-ui/${component}`}</Copyable>
         </GridItem>
       </GridContainer>
     </>
@@ -155,11 +170,13 @@ function DesktopInstall({ component, changelog, version }: InstallProps) {
 
 function CodeDocs({ component, readme, changelog }: BaseLayoutProps) {
   const viewport = useViewportSize();
-  const isMobile = viewport?.width < breakpoints.Tablet;
+  const isMobile = viewport?.width
+    ? viewport?.width < breakpoints.Tablet
+    : false;
 
-  const version = changelog.match(/(?<=<h2>)(.+?)(?=<\/h2>)/s)?.[1];
-  const example = readme.match(/(?<=js).*?(?=```)/s)?.[0];
-  const outputHTML = readme.match(/(?<=html).*?(?=```)/s)?.[0];
+  const version = changelog?.match(/(?<=<h2>)(.+?)(?=<\/h2>)/s)?.[1];
+  const example = readme?.match(/(?<=js).*?(?=```)/s)?.[0];
+  const outputHTML = readme?.match(/(?<=html).*?(?=```)/s)?.[0];
   const markdownAst = (unified()
     .use(markdown)
     .parse(readme) as unknown) as ReadmeMarkdown;
@@ -189,6 +206,7 @@ function CodeDocs({ component, readme, changelog }: BaseLayoutProps) {
                 </Code>
               </Tab>
             )}
+
             {outputHTML && (
               <Tab name="Output HTML" className={mt3} default={!example}>
                 <Code showLineNumbers language="xml">
@@ -202,6 +220,7 @@ function CodeDocs({ component, readme, changelog }: BaseLayoutProps) {
       <GridContainer align="flex-start" justify="flex-start">
         <GridItem sm={12} md={12} xl={12}>
           <PropTable markdownAst={markdownAst} component={component} />
+          <TypeDefinition markdownAst={markdownAst} readme={readme} />
         </GridItem>
       </GridContainer>
     </>