+++ website/styles/globals.tsx
@@ -1,37 +1,48 @@
 import { css } from '@emotion/core';
 
+const cdn = 'https://d2va9gm4j17fy9.cloudfront.net/fonts';
+
 export const globalStyles = css`
   @font-face {
     font-family: 'Akzidenz';
     font-weight: normal;
-    font-style: normal;
-    src: url('/fonts/akzidgrostdreg.eot');
-    src: url('/fonts/akzidgrostdreg.eot?#iefix') format('embedded-opentype'),
-      url('/fonts/akzidgrostdreg.woff') format('woff'),
-      url('/fonts/akzidgrostdreg.ttf') format('truetype'),
-      url('/fonts/akzidgrostdreg.svg#Akzidenz') format('svg');
+    src: url('${cdn}/akzidgrostdreg.eot');
+    /* IE9 Compat Modes */
+    src: url('${cdn}/akzidgrostdreg.eot?#iefix') format('embedded-opentype'),
+      /* IE6-IE8 */ url('${cdn}.akzidgrostdreg.woff') format('woff'),
+      /* Pretty Modern Browsers */ url('${cdn}/akzidgrostdreg.ttf')
+        format('truetype'),
+      /* Safari, Android, iOS */ url('${cdn}/akzidgrostdreg.svg#Akzidenz')
+        format('svg');
+    /* Legacy iOS */
   }
 
   @font-face {
     font-family: 'Akzidenz';
-    font-weight: medium;
-    font-style: normal;
-    src: url('fonts/akzidgrostdmed.eot');
-    src: url('fonts/akzidgrostdmed.eot?#iefix') format('embedded-opentype'),
-      url('fonts/akzidgrostdmed.woff') format('woff'),
-      url('fonts/akzidgrostdmed.ttf') format('truetype'),
-      url('fonts/akzidgrostdmed.svg#Akzidenz') format('svg');
+    font-weight: 600;
+    src: url('${cdn}/akzidgrostdmed.eot');
+    /* IE9 Compat Modes */
+    src: url('${cdn}/akzidgrostdmed.eot?#iefix') format('embedded-opentype'),
+      /* IE6-IE8 */ url('${cdn}/akzidgrostdmed.woff') format('woff'),
+      /* Pretty Modern Browsers */ url('${cdn}/akzidgrostdmed.ttf')
+        format('truetype'),
+      /* Safari, Android, iOS */ url('${cdn}/akzidgrostdmed.svg#Akzidenz')
+        format('svg');
+    /* Legacy iOS */
   }
 
   @font-face {
     font-family: 'Akzidenz';
     font-weight: bold;
-    font-style: normal;
-    src: url('fonts/akzidgrostdmed.eot');
-    src: url('fonts/akzidgrostdmed.eot?#iefix') format('embedded-opentype'),
-      url('fonts/akzidgrostdmed.woff') format('woff'),
-      url('fonts/akzidgrostdmed.ttf') format('truetype'),
-      url('fonts/akzidgrostdmed.svg#Akzidenz') format('svg');
+    src: url('${cdn}/akzidgrostdmed.eot');
+    /* IE9 Compat Modes */
+    src: url('${cdn}/akzidgrostdmed.eot?#iefix') format('embedded-opentype'),
+      /* IE6-IE8 */ url('${cdn}/akzidgrostdmed.woff') format('woff'),
+      /* Pretty Modern Browsers */ url('${cdn}/akzidgrostdmed.ttf')
+        format('truetype'),
+      /* Safari, Android, iOS */ url('${cdn}/akzidgrostdmed.svg#Akzidenz')
+        format('svg');
+    /* Legacy iOS */
   }
 
   html {