+++ packages/inline-definition/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @leafygreen-ui/inline-definition
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [ac0f3ff1]
+  - @leafygreen-ui/typography@7.0.0
+
 ## 2.0.1
 
 ### Patch Changes