+++ packages/pipeline/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @leafygreen-ui/pipeline
 
+## 2.1.1
+
+### Patch Changes
+
+- ee7923d3: Changes how we extend the types of HTMLElements, and standardizes how we document this across readmes
+- Updated dependencies [ee7923d3]
+- Updated dependencies [ee7923d3]
+  - @leafygreen-ui/tooltip@6.1.1
+  - @leafygreen-ui/lib@6.1.2
+
+## 2.1.0
+
+### Minor Changes
+
+- c18f16e6: Improves compatibility with React v17
+
+### Patch Changes
+
+- Updated dependencies [c18f16e6]
+- Updated dependencies [c18f16e6]
+- Updated dependencies [7359e552]
+- Updated dependencies [c9a0d89f]
+- Updated dependencies [9ee1d5fc]
+  - @leafygreen-ui/hooks@6.0.0
+  - @leafygreen-ui/tooltip@6.1.0
+  - @leafygreen-ui/palette@3.1.0
+  - @leafygreen-ui/lib@6.1.1
+
 ## 2.0.1
 
 ### Patch Changes